Dynamically adjust the position of the "Items" field on player cards
The "Items" field should be forced onto its own line if it would otherwise be the third field in its row, for the sake of text wrapping. It should *not* be forced onto a new line if it would otherwise only be the first or second field in its row.